Revert "Merge branch 'develop'"
This reverts commit de2d9f25bdf3e9f29bdaf3343cb04baa9d62d8f8, reversing
changes made to fa09412f94e38c110786d6542ab88bd455244651.

+++ src/main/java/com/mople/dto/response/pagination/FlatCursorPageResponse.java
@@ -3,11 +3,11 @@ package com.mople.dto.response.pagination;
 import java.util.List;
 
 public record FlatCursorPageResponse<T>(
-        Integer totalCount,
+        Long totalCount,
         List<T> content,
         CursorPage page
 ) {
-    public static <T> FlatCursorPageResponse<T> of(Integer totalCount, CursorPageResponse<T> response) {
+    public static <T> FlatCursorPageResponse<T> of(Long totalCount, CursorPageResponse<T> response) {
         return new FlatCursorPageResponse<>(
                 totalCount,
                 response.content(),
